updates to fitting tools
