+++ DrawHits/FitHistogram.py
@@ -1,0 +1,274 @@
+import ROOT
+
+class FitHistogram:
+
+    def interpolate(y,y1,y2,x1=0.,x2=1.):
+        ''' Linear interpolation between two points (x1,y1) and (x2,y2) 
+            to yield x corresponding to the target value y.
+        '''
+        return x1 + (y-y1)/(y2-y1)*(x2-x1)
+        
+    def __init__(self,histogram):
+        self.hist = histogram
+        self.graph_ = None
+        self.cumulativeGraph_ = None
+        self.cumNormGraph_ = None
+        self.cumNormTGraph_ = None
+        self.cumNormSpline_ = None               # spline corresponding to normalized cumulative graph
+
+    def graph(self):
+        ''' Histogram converted to list of (x,y) coordinates with x = bin center and y = bin contents.
+            Ignores under- / overflow.
+        '''
+        if self.graph_==None:
+            self.graph_ = [ ]
+
+            for i in range(self.hist.GetNbinsX()):
+                xbin = self.hist.GetBinLowEdge(i+1) + self.hist.GetBinWidth(i+1)/2.
+                self.graph_.append( (xbin,self.hist.GetBinContent(i+1)) )
+
+        return self.graph_
+
+    def cumulativeGraph(self):
+        ''' Histogram converted to list of (x,y) coordinates with x = bin center and y = cumulated bin contents.
+            Ignores under- / overflow.
+        '''
+        if self.cumulativeGraph_==None:
+            self.cumulativeGraph_ = [ ]
+            sum = 0.
+            for i in range(self.hist.GetNbinsX()):
+                xbin = self.hist.GetBinLowEdge(i+1) + self.hist.GetBinWidth(i+1)/2.
+                sum += self.hist.GetBinContent(i+1)
+                self.cumulativeGraph_.append( (xbin,sum) )
+
+        return self.cumulativeGraph_
+
+    def cumulativeNormGraph(self):
+        ''' Histogram converted to list of (x,y) coordinates with x = bin center and y = cumulated bin contents.
+            Normalized to total histogram contents (including under- / overflow).
+        '''
+        if self.cumNormGraph_==None:
+            cg = self.cumulativeGraph()
+            sum = self.hist.GetSumOfWeights()
+            self.cumNormGraph_ = [ ( x,y/sum ) for x,y in cg ]
+            
+        return self.cumNormGraph_
+
+    def cumulativeNormSpline(self):
+        ''' Create (TGraph and) TSpline3 from cumulativeNormGraph
+        '''
+        if self.cumNormTGraph_==None:
+            self.cumNormTGraph_ = ROOT.TGraph()
+            for x,y in self.cumulativeNormGraph():
+                self.cumNormTGraph_.SetPoint(self.cumNormTGraph_.GetN(),x,y)
+
+        if self.cumNormSpline_==None:
+            self.cumNormSpline_ = ROOT.TSpline3(self.hist.GetName()+"-spline",self.cumNormTGraph_)
+
+        return self.cumNormSpline_
+            
+        
+    def intersects(self,value,cumulative=False,norm=False,direction=0):
+        ''' Calculate x-coordinates for intersection(s) of a graph defined by a list of (x,y) points sorted in x
+              with y==value. Uses linear interpolation.
+            Arguments:
+               value ....... target value
+               cumulative .. if true, use cumulative graph
+               norm ........ if true, use normalized cumulative graph
+               direction ... three possible values: 0 = any intersection, +1/-1 = consider only segments
+                             with positive / negative slope.
+        '''
+        #
+        # Get graph and do basic check
+        #
+        graph = None
+        if cumulative:
+            graph = self.cumulativeNormGraph() if norm else self.cumulativeGraph()
+        else:
+            assert norm==False
+            graph = self.graph()
+            
+        result = [ ]
+        if len(graph)<2:
+            return result
+        #
+        # loop over adjacent pairs of points
+        #
+        x1 = None
+        y1 = None
+        for x2,y2 in graph:
+            #
+            # start checking at 2nd point
+            #
+            if x1!=None:
+                assert x2>x1
+                #
+                # value in interval?
+                #
+                if value>=min(y1,y2) and value<=max(y1,y2):
+                    # check if dy is positive or negative, and compare with required sign of direction
+                    if direction==0 or direction*(y2-y1)>0:
+                        result.append(FitHistogram.interpolate(value,y2,y1,x2,x1))
+            #
+            # move to next point
+            #
+            x1 = x2
+            y1 = y2
+
+        return result
+
+    def fwhm(self):
+        ''' Return lowest / highest x corresponding to ymax/2, and ymax/2
+        '''
+        #
+        # target y value (1/2 maximum)
+        #
+        y = self.hist.GetBinContent(self.hist.GetMaximumBin())/2.
+        #
+        # use first intersection with upward slope
+        #
+        xups = self.intersects(y,cumulative=False,direction=1)
+        print("xups",xups)
+        xlow = xups[0] if xups else None
+        #
+        # use last intersection with downward slope
+        #
+        xdowns = self.intersects(y,cumulative=False,direction=-1)
+        print("xdowns",xdowns)
+        xhigh = xdowns[-1] if xdowns else None
+        #dx = self.hist.GetBinWidth(1)
+        ###
+        ## preset result (low / high x values)
+        ##
+        #xlow = None
+        #xhigh = None
+        ##
+        ## loop over pairs of adjacent histogram bins
+        ##
+        #x1 = self.hist.GetBinLowEdge(1)
+        #y1 = self.hist.GetBinContent(1)
+        #for i in range(2,self.hist.GetNbinsX()):
+        #    # check if target value between contents of neighbouring bins
+        #    x2 = self.hist.GetBinLowEdge(i)
+        #    y2 = self.hist.GetBinContent(i)
+        #    first = None
+        #    if y>=y1 and y<y2:
+        #        first = True
+        #    elif y<=y1 and y>y2:
+        #        first = False
+        #
+        #    if first!=None:
+        #        # calculate interpolated x value
+        #        x = FitHistogram.interpolate(y,y1,y2,x1,x2)
+        #        # store lowest and highest x corresponding to y
+        #        if first and xlow==None:
+        #            xlow = x
+        #        elif not first:
+        #            xhigh = x
+        #    # move to next bin
+        #    x1 = x2
+        #    y1 = y2
+
+        #
+        # require result ( assumes that first and last bins are < ymax/2 ) and
+        # correct for bin width / 2 ( assumes that bin value corresponds to center of bin )
+        assert xlow!=None and xhigh!=None
+        return (xlow,xhigh,y)
+        #return (xlow+dx,xhigh+dx,y)
+
+    def quantile(self,prob):
+        ''' Return x-value to quantile q. 
+        '''
+        result = self.intersects(prob,cumulative=True,norm=True,direction=1)
+        #print(prob,result)
+        assert len(result)<2
+
+        return result[0] if result else None
+
+    def findRootSpline(self,value,eps=0.001):
+        ''' Find position where spline derived from cumulative NormGraph= value. Assumes that the spline is \
+            monotonously increasing. Tolerance is eps*<difference between first and last point)
+        '''
+        #
+        # Make sure spline is available
+        #
+        spline = self.cumulativeNormSpline()
+        #
+        # No result if first / last point is above / below required value
+        #
+        cGraph = self.cumulativeGraph()
+        if len(cGraph)==0 or cGraph[0][1]>value or cGraph[-1][1]<value:
+            return None
+        #
+        # Initialize from first and last point and verify that monotonicity
+        #
+        xl = spline.GetXmin()
+        yl = spline.Eval(xl)
+        xh = spline.GetXmax()
+        yh = spline.Eval(xh)
+        ymin = yl
+        ymax = yh
+        if ymin>=ymax:
+            print("findRootSpline: last value <= first value")
+            return None
+        #
+        # Check if inside values spanned by spline
+        #
+        if value<ymin or value>ymax:
+            print("findRootSpline: required value outside range")
+            return None
+        #
+        # tolerance for distance between points
+        #
+        dxmax = eps*(ymax-ymin)
+        #
+        # loop with cutoff in case of non-convergence
+        #
+        found = False
+        for i in range(1000):
+            if (xh-xl)<dxmax:
+                found = True
+                break
+            x = (xl+xh)/2.
+            y = spline.Eval(x)
+            if y<=value:
+                xl = x
+                yl = y
+            else:
+                xh = x
+                yh = y
+        if not found:
+            print("findRootSpline: did not converge")
+            return None
+
+        return (xl+xh)/2.
+            
+    def fitGaus(self,prob1=0.,prob2=1.):
+        assert prob2>prob1
+        result = ( None, None )
+        xmin = self.hist.GetXaxis().GetXmin()
+        if prob1>0.:
+            xmin = self.findRootSpline(prob1)
+            if xmin==None:
+                return result
+        xmax = self.hist.GetXaxis().GetXmax()
+        if prob2<1.:
+            xmax = self.findRootSpline(prob2)
+            if xmin==None:
+                return result
+
+        fitFunc = ROOT.TF1("mygaus","gaus(0)",xmin,xmax)
+        fitFunc.SetParameter(0,self.hist.GetMaximum())
+        fitFunc.SetParameter(1,0.)
+        fitFunc.SetParLimits(1,-10.,10.)
+        fitFunc.SetParameter(2,self.hist.GetRMS())
+        
+        fitPtr = self.hist.Fit(fitFunc,"S0")
+        #fitPtr = self.hist.Fit("gaus","S0","",xmin,xmax)
+        if ( not fitPtr.IsValid() ) or fitPtr.IsEmpty():
+            return result
+
+        #return fitPtr,self.hist.GetFunction("gaus")
+        return fitPtr,fitFunc
+            
+
